requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
chainlit
langchain
langchain_community
PyPDF2
chromadb
groq
langchain-groq
